refactor: test mocks into CRUD & socket configs
- create new test configuration repositories `CRUD` and `socket` in `test\mocks\config` repository within root of the project.
- move `mockSocket.js` and `mockSocketModel.js` to `test\mocks\config\socket` repository and update the import path instances across the project.
- move `mockRes.js` and `mockModel.js` to `test\mocks\config\CRUD` repository and update the import path instances across the project.
