+++ test/mocks/config/socket/mockSocket.js
