+++ test/unit/backend/controller/chatUnit.test.js
@@ -1,6 +1,6 @@
 import { expect } from "chai";
-import MockModel from "../../../mocks/config/mockModel.js";
-import MockRes from "../../../mocks/config/mockRes.js";
+import MockModel from "../../../mocks/config/CRUD/mockModel.js";
+import MockRes from "../../../mocks/config/CRUD/mockRes.js";
 import MockData from "../../../mocks/config/mockData.js";
 import {
   getMessages,
