+++ test/unit/backend/controller/classroomUnit.test.js
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import MockData from "../../../mocks/config/mockData.js";
-import MockModel from "../../../mocks/config/mockModel.js";
-import MockRes from "../../../mocks/config/mockRes.js";
+import MockModel from "../../../mocks/config/CRUD/mockModel.js";
+import MockRes from "../../../mocks/config/CRUD/mockRes.js";
 
 import {
   joinClassroom,
