+++ test/unit/backend/controller/examUnit.test.js
@@ -1,7 +1,7 @@
 import { expect } from "chai";
-import MockModel from "../../../mocks/config/mockModel.js";
+import MockModel from "../../../mocks/config/CRUD/mockModel.js";
 import MockData from "../../../mocks/config/mockData.js";
-import MockRes from "../../../mocks/config/mockRes.js";
+import MockRes from "../../../mocks/config/CRUD/mockRes.js";
 import {
   getExam,
   getExamFeedback,
